PR Feedback

+++ OrderCloud.Integrations.Payment.CardConnect/Mappers/CardConnectFundReversalResponseMapper.cs
@@ -16,7 +16,7 @@ namespace OrderCloud.Integrations.Payment.CardConnect.Mappers
 				ResponseCode = transaction.respcode,
 				Message = transaction.resptext,
 				TransactionID = transaction.retref,
-				Amount = Convert.ToInt64(transaction.amount),
+				Amount = Convert.ToDecimal(transaction.amount),
             };
 		}
 	}
